requirements.txt
openai==1.3.5
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv